+++ widget/release_notes.md
@@ -1,2 +1,3 @@
-# What's New
-Minor bug fix
+## What's New
+
+- This widget can now be configured to listen to events and dynamically load data broadcasted by another widget in the same dashboard.